mcp/pyproject.toml
[project]
name = "corvus-mcp"
version = "0.1.0"
description = "MCP server exposing Corvus academic search tools (Semantic Scholar + citation graph)"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
